Added add and remove property

+++ Engine/Character.cpp
@@ -4,7 +4,7 @@
 
 void Character::Draw(Graphics & gfx, int dimension) const
 {
-	SpriteCodex::DrawCharacter(pos.x*dimension, pos.y*dimension, gfx);
+	SpriteCodex::DrawCharacter(int(pos.x*dimension), int(pos.y*dimension), gfx);
 }
 
 void Character::Move(const Vec2& dir, bool canGo)
